final commit
1: accountManager Component
Create the accountManager component to retrieve the data using the AccountManagerController class. Use the imperative method in the connectedCallback and display the accounts data in a lightning-datatable with at least five fields (set keyField="Id").

Assignment 2: accountManagerFetch Component
Create the accountManagerFetch component to retrieve account data using the wire method instead of the imperative method. Display the accounts data in a lightning-datatable with at least five fields (set keyField="Id").